requirements.txt
flask[async]==2.3.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
anthropic==0.18.1
flask-wtf==1.2.1
wtforms==3.0.1
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
